refactor(server): move aiQuota.ts and tests to apps/server/src/modules/
